Refactor Code || Add TODO list || Tidy Up Code || Tidy Up Unity Assets
I have grossly neglected the maintenance of this code and will now make up for it. Every script has been updated to make the code easier to understand with comments, refactoring, and repositioning of serialized fields, cashed references, and attributes. Each script also has a block comment explaining what the script is attached to and what it does.